Add support for sending user-defined encrypted to-device messages
This is a port of the same change from the robertlong/group-call branch.
